+++ drift_dev/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2-dev
+
+- Also transform `analysis_options.yaml` files in the `drift_dev migrate` command.
+
 ## 1.0.1
 
 This is the initial release of the `drift_dev` package (formally known as `moor_generator`).